+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.1.0
+## 0.1.1
 
-- Initial Release
+- Port to 1.21
+- Fix client crash when connecting to dedicated server.
+- Fix privileges not being cleared on reload.